Refactor ops out of undo stack
