+++ src/renderer/editor/clipboard.ts
@@ -3,7 +3,8 @@ import { getSelectionRanges } from './selection';
 import { clipboardUpdated } from '../store/clipboard/actions';
 import store from '../store/store';
 import dispatchOp from '../store/dispatchOp';
-import { makeDeleteSelection, makePasteWord } from '../store/undoStack/ops';
+import { makeDeleteSelection } from '../store/transcriptionWords/ops/deleteSelection';
+import { makePasteWord } from '../store/transcriptionWords/ops/pasteWord';
 
 const { dispatch } = store;
