+++ src/renderer/store/transcriptionWords/reducer.ts
@@ -2,8 +2,8 @@ import { Reducer } from 'react';
 import { mapInRanges } from 'renderer/utils/list';
 import { Word } from 'sharedTypes';
 import { Action } from '../action';
-import { mergeWords } from './mergeWords';
-import { splitWord } from './splitWord';
+import { mergeWords } from './helpers/mergeWordsHelper';
+import { splitWord } from './helpers/splitWordHelper';
 import {
   DELETE_SELECTION,
   MERGE_WORDS,
